relocate to client from common
