+++ client/src/main/java/dev/qf/client/CartController.java
@@ -1,4 +1,7 @@
-package common;
+package dev.qf.client;
+
+import common.Cart;
+import common.OrderItem;
 
 public class CartController {
     private final Cart cart;
